requirements/common.txt
Django==4.1.7
gunicorn==20.1.0
ortools==9.5.2237
pika==1.3.1
python-decouple==3.8
